% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkDate}
\alias{checkDate}
\title{Documentation}
\usage{
checkDate(date)
}
\arguments{
\item{date}{The string we wish to check if it's a proper date format}
}
\value{
boolean to indicate if date is in the format of YYYY-MM-DD
}
\description{
Checks the input date if it's proper format.
}
\details{
This functions return TRUE if the date for format
}
\note{
When \code{date} is ignored it will return FALSE
}
\examples{
checkDate("")
checkDate("2014-11-17") # return TRUE
checkDate("2014November17th") # return FALSE
}

