% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkNHLConnection}
\alias{checkNHLConnection}
\title{Documentation}
\usage{
checkNHLConnection()
}
\value{
boolean to indicate if connection to `http://live.nhle.com/` server is good
}
\description{
Checks if the nhl server is available for query
}
\details{
This functions return TRUE if the connection is good.
}
\note{
If it's false, it means either the internet connection is not available or
the NHL server is down... though not likely cuz they can't afford that downtime!
(remember to pay your internet bill)
}
\examples{
checkNHLConnection()
}

