% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Documentation}
\usage{
gday(team.name = "canucks")
}
\arguments{
\item{team.name}{The name of your team}
}
\value{
logical \code{TRUE} if \code{team.name} has a NHL game today
\code{FALSE} otherwise
}
\description{
Is it Gameday?
}
\details{
This function returns TRUE if your NHL team plays today
and FALSE otherwise

Now here's the input
}
\note{
case in \code{team.name} is ignored it will check for Canucks
}
\examples{
gday("canucks")
gday("Bruins")
}

