% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score}
\alias{score}
\title{Documentation}
\usage{
score(date = Sys.Date(), team = NULL)
}
\arguments{
\item{date}{The scores of the date you want to search for. format: YYYY-MM-DD}

\item{team}{The team of your interest, can the team's city name, too.}
}
\value{
data.frame If connection to live.nhle.com is successful and
that the date is the right format
}
\description{
Gives a summary of the scores from all teams playing on the given date.
Default date is that of the current day.
}
\details{
This functions return a data.frame of the scores of all home teams and away teams,
as available on live.nhle.com.
}
\note{
When \code{date} is ignored it will check for \code{Sys.Date()},
and if \code{team} is ignored, the function will return all the teams playing on
the given day.
If the input team playing, the function will return the score information of both the team and its rival team.
If the input team is not playing, the function prints "The team isn't playing today"
and returns a data frame of all the games going on today.
The function also throw errors when date format is not proper or
information are not avaiable for certain date.
\code{team} can be either the name of the team or the city names
}
\examples{
score()
score(date = "2014-11-17")
score(team = "kings")
score(date = "2014-11-17", team = "kings")
}

