% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkDate}
\alias{checkDate}
\title{Documentation}
\usage{
checkDate(date = Sys.Date())
}
\arguments{
\item{date}{The string we wish to check if it's a proper date format}
}
\value{
boolean to indicate if date is in the format of YYYY-MM-DD
}
\description{
Checks the input date if it's proper format.
}
\details{
This functions return TRUE if the date for format, otherwise FALSE
}

