% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Documentation}
\usage{
gday(date = Sys.Date(), team = "canucks")
}
\arguments{
\item{team}{The name of your team or the city of the team.}
}
\value{
string Name of the channel to watch. Return "information not available for this date"
if the NHL server cannot provide information for the given date.
Returns "not available" if the team is #' not playing on the current day,
"undecided" if the team is playing but the channel name #' is not available on nhl
}
\description{
Is it the game day of your favorite team?
}
\details{
This function will tell you which channel to watch if your team is playing,
otherwise it tells you to wait for a few more days

Now here's the input
}
\note{
case in \code{team} is ignored it will check for Canucks because this package
was written by someone in Vancouver
}
\examples{
gday() # check for "Canucks"
}

