% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gday}
\alias{gday}
\title{Documentation}
\usage{
gday(team = "canucks")
}
\arguments{
\item{team}{The name of your team or the city of the team.}
}
\value{
string Name of the channel to watch
}
\description{
Is it the game day of your favorite team?
}
\details{
This function will tell you which channel to watch if your team is playing,
otherwise it tells you to wait for a few more days

Now here's the input
}
\note{
case in \code{team} is ignored it will check for Canucks because this package
was written by someone in Vancouver
}
\examples{
gday() # check for "Canucks"
gday("Bruins")
}

